update noble integer
Earlier solution was not working for test cases like {-4, -2, 0, -1, -6} which has 0 as max value and also nos greater than 0 are 0. But after sorting when the loop iterates, it may sometimes skip 0 when the digit after array is 0( as A[i]==A[i+1] continue , skips this 0 ). You may check this on interviewbit.